+++ dtwc/main.cpp
@@ -18,7 +18,7 @@ int main()
   int Ndata_max = 20; // Load maximum 20 of data.
   // Note: Run this from the project root directory, or specify an absolute path
   dtwc::DataLoader dl{ std::filesystem::path("data") / "dummy", Ndata_max };
-  dl.startColumn(1).startRow(1); // Since dummy files are in Pandas format skip first row/column.
+  dl.start_column(1).start_row(1); // Since dummy files are in Pandas format skip first row/column.
 
   dtwc::Problem prob{ probName, dl }; // Create a problem.
   prob.maxIter = 100;