+++ tests/unit/unit_test_distance_matrix_properties.cpp
@@ -24,7 +24,7 @@
 #endif
 
 static struct TestDataInit2 {
-  TestDataInit2() { dtwc::settings::paths::setDataPath(DTWC_TEST_DATA_DIR); }
+  TestDataInit2() { dtwc::settings::paths::set_data_path(DTWC_TEST_DATA_DIR); }
 } test_data_init2_;
 
 using Catch::Matchers::WithinAbs;
@@ -38,7 +38,7 @@ namespace {
 Problem make_problem(int N_data)
 {
   dtwc::DataLoader dl{ settings::paths::data / "dummy", N_data };
-  dl.startColumn(1).startRow(1);
+  dl.start_column(1).start_row(1);
   dtwc::Problem prob{ "dist_mat_test", dl };
   return prob;
 }